changes done as to review

+++ server/www/global-filters.js
@@ -1,13 +1,17 @@
 const API = require('../utils/api');
 
-exports.list = class extends API {
+class Global_filters extends API {
 
 	async list() {
-		return await this.mysql.query(`SELECT * FROM tb_global_filters WHERE account_id=?`, [this.account.account_id]);
-	}
-};
+		const result = await this.mysql.query(`SELECT * FROM tb_global_filters WHERE account_id = ?`, [this.account.account_id]);
 
-exports.insert = class extends API {
+		for(const data of result) {
+			data.placeholder = data.placeholder.split(',');
+		}
+
+		return result;
+	};
+
 	async insert() {
 		this.user.privilege.needs('administrator');
 
@@ -21,20 +25,18 @@ exports.insert = class extends API {
 			offset: this.request.body.offset || null,
 			dataset: this.request.body.dataset,
 		};
+
 		return await this.mysql.query(
 			`INSERT INTO tb_global_filters SET ?`,
 			[params],
 			'write'
 		);
 	}
-}
 
-exports.update = class extends API {
 	async update() {
 		this.user.privilege.needs('administrator');
 
 		const params = {
-			account_id: this.account.account_id,
 			name: this.request.body.name,
 			placeholder: this.request.body.placeholder,
 			default_value: this.request.body.default_value,
@@ -45,21 +47,24 @@ exports.update = class extends API {
 		};
 
 		return await this.mysql.query(
-			`UPDATE tb_global_filters SET ? WHERE id = ?`,
-			[params, this.request.body.id],
+			`UPDATE tb_global_filters SET ? WHERE id = ? and account_id = ?`,
+			[params, this.request.body.id, this.account.account_id],
 			'write'
 		);
 	}
-}
 
-exports.delete = class extends API {
 	async delete() {
 		this.user.privilege.needs('administrator');
 
 		return await this.mysql.query(
-			`DELETE FROM tb_global_filters WHERE id = ?`,
-			[this.request.body.id],
+			`DELETE FROM tb_global_filters WHERE id = ? and account_id = ?`,
+			[this.request.body.id, this.account.account_id],
 			'write'
 		);
 	}
-}
+};
+
+exports.list = Global_filters;
+exports.insert = Global_filters;
+exports.update = Global_filters;
+exports.delete = Global_filters;
